requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
